# Conflicts:
#	app/src/main/java/com/terning/point/di/AuthInterceptor.kt
#	app/src/main/java/com/terning/point/di/DataSourceModule.kt
#	app/src/main/java/com/terning/point/di/RepositoryModule.kt
#	app/src/main/java/com/terning/point/di/ServiceModule.kt
#	data/src/main/java/com/terning/data/service/InternService.kt
#	feature/src/main/java/com/terning/feature/intern/InternRoute.kt
#	feature/src/main/java/com/terning/feature/intern/InternViewModel.kt
#	feature/src/main/java/com/terning/feature/main/MainNavigator.kt
#	feature/src/main/java/com/terning/feature/search/search/component/SearchIntern.kt

+++ data/src/main/java/com/terning/data/repositoryimpl/InternRepositoryImpl.kt
@@ -8,7 +8,7 @@ import javax.inject.Inject
 class InternRepositoryImpl @Inject constructor(
     private val internDataSource: InternDataSource,
 ) : InternRepository {
-    override suspend fun getInternInfo(id: Int): Result<InternInfoModel> =
+    override suspend fun getInternInfo(id: Long): Result<InternInfoModel> =
         runCatching {
             internDataSource.getInternInfo(id).result.toInternEntity()
         }
